Renamed the controller named 'cards' to the cardsController
Updated imports for that controller
